backend/requirements.txt
fastapi
strawberry-graphql[fastapi]
strawberry
sqlalchemy
pandas
python-dateutil
email-validator
phonenumbers
